Project page setup

+++ app/src/main/java/com/openclassrooms/entrevoisins/ui/neighbour_list/NeighbourFragment.java
@@ -30,20 +30,18 @@ import java.util.List;
 public class NeighbourFragment extends Fragment {
 
     private NeighbourApiService mApiService;
-    private List<Neighbour> mNeighbours;
     private RecyclerView mRecyclerView;
-    //TODO : J'ai ajouté une variable favori pour qu'on sache si le fragment est celui des voisins ou des favoris
     private boolean isFavorite;
 
     public static String FAVORITE_KEY = "FAVORITE_KEY";
 
     /**
      * Create and return a new instance
+     *
      * @return @{@link NeighbourFragment}
      */
     public static NeighbourFragment newInstance(boolean isFavorite) {
         NeighbourFragment fragment = new NeighbourFragment();
-        //TODO J'ai ajouté ce code pour qu'on puisse instancier le fragment avec l'info de favori
         Bundle bundle = new Bundle();
         bundle.putBoolean(FAVORITE_KEY, isFavorite);
         fragment.setArguments(bundle);
@@ -61,7 +59,6 @@ public class NeighbourFragment extends Fragment {
                              Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_neighbour_list, container, false);
         Context context = view.getContext();
-        //TODO : C'est ici qu'on récupère l'information favoris
         assert getArguments() != null;
         isFavorite = getArguments().getBoolean(FAVORITE_KEY, false);
         mRecyclerView = (RecyclerView) view;
@@ -74,12 +71,13 @@ public class NeighbourFragment extends Fragment {
      * Init the List of neighbours
      */
     private void initList() {
-        if(isFavorite){
-            mNeighbours = mApiService.getFavoriteNeighbours();
-        }else {
-            mNeighbours = mApiService.getNeighbours();
+        List<Neighbour> neighboursList;
+        if (isFavorite) {
+            neighboursList = mApiService.getFavoriteNeighbours();
+        } else {
+            neighboursList = mApiService.getNeighbours();
         }
-        mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(mNeighbours, isFavorite));
+        mRecyclerView.setAdapter(new MyNeighbourRecyclerViewAdapter(neighboursList, isFavorite));
     }
 
     @Override
@@ -102,57 +100,58 @@ public class NeighbourFragment extends Fragment {
 
     /**
      * Fired if the user clicks on a delete button
+     *
      * @param event : neighbour
      */
     @Subscribe
     public void onDeleteNeighbour(DeleteNeighbourEvent event) {
-        mApiService.deleteNeighbour(event.neighbour);
-        mApiService.deleteFavoriteNeighbour(event.neighbour);
-        initList();
+        if (!isFavorite) {
+            mApiService.deleteNeighbour(event.neighbour);
+            mApiService.deleteFavoriteNeighbour(event.neighbour);
+            initList();
+        }
     }
 
-    //TODO : Là on mimique la réception du click sur un neighbour
     /**
      * Fired if the user clicks on cell
+     *
      * @param event : neighbour
      */
     @Subscribe
     public void onClickNeighbour(ClickNeighbourEvent event) {
-        launchDetailActivity(event.neighbour);
+        if (!isFavorite)
+            launchDetailActivity(event.neighbour);
     }
-
-    //TODO : A FAIRE -> Faire la mimique de la réception des delete pour un favorite neighbour (ok)
 
     /**
      * In the page favorite Fired if the user clicks on a delete button
+     *
      * @param event : favorite neighbour
      */
     @Subscribe
     public void onDeletefavoriteNeighbour(DeleteFavoriteNeighbourEvent event) {
-    //TODO : A FAIRE -> Dans la fonction delete, appeler la fonction de l'apiService pour supprimer le favorite neighbour (ok)
-
+        if (isFavorite) {
             mApiService.deleteFavoriteNeighbour(event.neighbour);
             initList();
-
+        }
     }
 
-    //TODO : A FAIRE -> Faire la mimique de la réception des clicks  pour un favorite neighbour (ok)
     /**
      * In the page favorite  Fired if the user clicks on cell
+     *
      * @param event : favorite neighbour
      */
     @Subscribe
     public void onClickFavoriteNeighbour(ClickFavoriteNeighbourEvent event) {
-    //TODO : A FAIRE -> Dans la fonction click, lancer l'activité avec la valeur favorite true(ok)
-
+        if (isFavorite)
             launchDetailActivity(event.neighbour);
-
     }
 
     private void launchDetailActivity(Neighbour neighbour) {
         Intent intent = new Intent(getActivity(), NeighbourDetailsActivity.class);
         intent.putExtra(NeighbourDetailsActivity.NEIGHBOUR_KEY, neighbour);
-        intent.putExtra(NeighbourDetailsActivity.FAVORITENEIGHBOUR_KEY, mApiService.isNeighbourFavorite(neighbour));
+        intent.putExtra(NeighbourDetailsActivity.FAVORITE_NEIGHBOUR_KEY, mApiService
+                .isNeighbourFavorite(neighbour));
         startActivity(intent);
     }
